refactor(ui): drop unnecessary 'use client' from base-ui wrappers
Follows shadcn-ui/ui#10167 — base-ui v1.2+ declares its own client
boundary, so wrapper components don't need 'use client'. Dropping the
directive lets server components import helpers like buttonVariants
without the 'called from the server' error.

Kept 'use client' only where the wrapper itself uses React hooks:
calendar, field, toaster.

Reverts the buttonVariants split (button-variants.ts) now that the
original single-file shape works from server components.
