+++ packages/ui/src/components/avatar.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Avatar as AvatarPrimitive } from "@base-ui/react/avatar";
 import * as React from "react";
