+++ packages/ui/src/components/popover.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Popover as PopoverPrimitive } from "@base-ui/react/popover";
 import * as React from "react";
