+++ packages/ui/src/components/progress.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Progress as ProgressPrimitive } from "@base-ui/react/progress";
 
 import { cn } from "../lib/utils";
