+++ packages/ui/src/components/switch.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Switch as SwitchPrimitive } from "@base-ui/react/switch";
 
 import { cn } from "../lib/utils";
