+++ packages/ui/src/components/table.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import type { ComponentProps } from "react";
 
 import { cn } from "../lib/utils";
